Refactor transaction out.
